+++ examples/resources/closure-js/externs/lucuma-examples-externs.js
@@ -1,6 +1,0 @@
-var Ladda = {};
-Ladda.create = function(el) {};
-Ladda.stopAll = function() {};
-
-var Prism = {};
-Prism.prototype.highlightElement = function(el) {};